Refactor permissions middleware to move redundant functionality into middleware factory function.
